Add functionality to save configuration to a TOML file
This update introduces a new option for saving the current app configuration into a TOML file. The upgrade affects ArgParser in the main function by adding the option "save-config". Also, a helper method "saveToToml" has been added to handle the actual file writing process. The Ktoml library is employed to perform the encoding of the configuration object.